requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.27
langchain-tavily==0.2.4
langchain-text-splitters==0.3.8
langgraph==0.5.0
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.1
langgraph-sdk==0.1.72
langgraph-supervisor==0.0.27
langsmith==0.4.4
marshmallow==3.26.1
multidict==6.6.1
mypy==1.16.1
mypy-extensions==1.1.0
numpy==2.2.6
openai==1.93.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pathspec==0.12.1
propcache==0.3.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlalchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
